proper output; moved auxiliary functions to utils.py
